// frontend/web/parceldelivery-admin-app/src/app/_dtos/vehicle-dto.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export interface VehicleDto {
    id: string | undefined;
    registrationNumber: string | undefined;
    type: string | undefined;
    year: number | undefined;
    technicalInspectionExpirationDate: Date | undefined;
    seatingCapacity: number | undefined;
    maxInternalSpaceX: number | undefined;
    maxInternalSpaceY: number | undefined;
    maxInternalSpaceZ: number | undefined;
}